requirements.txt
fastapi
uvicorn
pandas
pydantic
streamlit
python-dotenv
requests
torch
transformers
scipy